+++ web-client/src/main/java/no/nav/familie/webflux/builder/AzureWebClientConfig.kt
@@ -1,9 +1,9 @@
 package no.nav.familie.webflux.builder
 
+import no.nav.familie.webflux.filter.BearerTokenFilterFunction
 import no.nav.familie.webflux.filter.ConsumerIdFilterFunction
 import no.nav.familie.webflux.filter.InternLoggerFilterFunction
 import no.nav.familie.webflux.filter.MdcValuesPropagatingFilterFunction
-import no.nav.familie.webflux.filter.StsBearerTokenFilterFunction
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
@@ -14,19 +14,17 @@ import org.springframework.web.reactive.function.client.WebClient
 @Configuration
 @Import(ConsumerIdFilterFunction::class,
         InternLoggerFilterFunction::class,
-        StsBearerTokenFilterFunction::class)
-class StsWebClientBuilder {
+        BearerTokenFilterFunction::class)
+class AzureWebClientConfig {
 
-    @Bean("stsWebClientBuilder")
-    fun stsWebClientBuilder(consumerIdFilterFunction: ConsumerIdFilterFunction,
-                            internLoggerFilterFunction: InternLoggerFilterFunction,
-                            stsBearerTokenFilterFunction: StsBearerTokenFilterFunction): WebClient.Builder {
+    @Bean("azureWebClientBuilder")
+    fun azureWebClientBuilder(consumerIdFilterFunction: ConsumerIdFilterFunction,
+                              internLoggerFilterFunction: InternLoggerFilterFunction,
+                              bearerTokenFilterFunction: BearerTokenFilterFunction): WebClient.Builder {
         return WebClient.builder()
                 .filter(consumerIdFilterFunction)
-                .filter(internLoggerFilterFunction)
-                .filter(stsBearerTokenFilterFunction)
+                .filter(bearerTokenFilterFunction)
                 .filter(MdcValuesPropagatingFilterFunction())
     }
-
 }
